feat: add subscription to AddPolylineRequested in MapBloc
Refactor CreatePolyine in map event change field with to double to int

+++ lib/features/home/presentation/bloc/map/map_event.dart
@@ -85,14 +85,14 @@ class CreatePolyline extends MapEvent {
   final String polylineId;
   final List<LatLng> points;
   final Color color;
-  final double width;
+  final int width;
   final List<PatternItem>? patterns;
 
   const CreatePolyline({
     this.polylineId = 'route',
     required this.points,
     this.color = const Color(0xFF2196F3),
-    this.width = 5.0,
+    this.width = 5,
     this.patterns = const [],
   });
